chore: move DynamicSelectTextField to feature-compose/feature-component
