+++ feature-compose/feature-settings/src/main/kotlin/org/gampiot/robok/feature/settings/compose/ui/screens/rdkmanager/ConfigureRDKScreen.kt
@@ -15,6 +15,7 @@ import org.koin.core.parameter.parametersOf
 import org.gampiot.robok.feature.settings.compose.viewmodels.ConfigureRDKViewModel
 import org.gampiot.robok.feature.component.compose.preferences.base.PreferenceLayout
 import org.gampiot.robok.feature.component.compose.preferences.base.PreferenceGroup
+import org.gampiot.robok.feature.component.compose.textfields.DynamicSelectTextField
 import org.gampiot.robok.feature.res.Strings
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -55,50 +56,6 @@ fun ConfigureRDKScreen(
                 }
             ) {
                 Text(text = stringResource(id = Strings.common_word_save))
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun DynamicSelectTextField(
-    selectedValue: String,
-    options: List<String>,
-    label: String,
-    onValueChangedEvent: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    var expanded by remember { mutableStateOf(false) }
-
-    ExposedDropdownMenuBox(
-        expanded = expanded,
-        onExpandedChange = { expanded = !expanded },
-        modifier = modifier
-    ) {
-        OutlinedTextField(
-            readOnly = true,
-            value = selectedValue,
-            onValueChange = {},
-            label = { Text(text = label) },
-            trailingIcon = {
-                ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded)
-            },
-            colors = OutlinedTextFieldDefaults.colors(),
-            modifier = Modifier
-                .menuAnchor()
-                .fillMaxWidth()
-        )
-
-        ExposedDropdownMenu(expanded = expanded, onDismissRequest = { expanded = false }) {
-            options.forEach { option: String ->
-                DropdownMenuItem(
-                    text = { Text(text = option) },
-                    onClick = {
-                        expanded = false
-                        onValueChangedEvent(option)
-                    }
-                )
             }
         }
     }
